+++ example/test/sample_test.dart
@@ -15,9 +15,11 @@ void main() {
     Future<void> bddSetUp(WidgetTester tester) async {
       await theAppIsRunning(tester);
     }
+
     Future<void> bddTearDown(WidgetTester tester) async {
       await iDoNotSeeText(tester, 'surprise');
     }
+
     testWidgets('''Initial counter value is 0''', (tester) async {
       try {
         await bddSetUp(tester);
@@ -35,7 +37,8 @@ void main() {
         await bddTearDown(tester);
       }
     });
-    testWidgets('''Outline: Plus button increases the counter (0, '0')''', (tester) async {
+    testWidgets('''Outline: Plus button increases the counter (0, '0')''',
+        (tester) async {
       try {
         await bddSetUp(tester);
         await theAppIsRunning(tester);
@@ -45,7 +48,8 @@ void main() {
         await bddTearDown(tester);
       }
     });
-    testWidgets('''Outline: Plus button increases the counter (1, '1')''', (tester) async {
+    testWidgets('''Outline: Plus button increases the counter (1, '1')''',
+        (tester) async {
       try {
         await bddSetUp(tester);
         await theAppIsRunning(tester);
@@ -55,7 +59,8 @@ void main() {
         await bddTearDown(tester);
       }
     });
-    testWidgets('''Outline: Plus button increases the counter (42, '42')''', (tester) async {
+    testWidgets('''Outline: Plus button increases the counter (42, '42')''',
+        (tester) async {
       try {
         await bddSetUp(tester);
         await theAppIsRunning(tester);