+++ lib/src/step_file.dart
@@ -13,8 +13,18 @@ abstract class StepFile {
     String line,
     Map<String, String> existingSteps,
     GeneratorOptions generatorOptions,
+    String testerTypeTagValue,
+    String testerNameTagValue,
   ) {
     final file = '${getStepFilename(line)}.dart';
+
+    final testerType = testerTypeTagValue.isNotEmpty
+        ? testerTypeTagValue
+        : generatorOptions.testerType;
+
+    final testerName = testerNameTagValue.isNotEmpty
+        ? testerNameTagValue
+        : generatorOptions.testerName;
 
     if (existingSteps.containsKey(file)) {
       final import =
@@ -33,7 +43,14 @@ abstract class StepFile {
       final import =
           p.join(generatorOptions.stepFolder, file).replaceAll(r'\', '/');
       final filename = p.join(featureDir, generatorOptions.stepFolder, file);
-      return NewStepFile._(import, filename, package, line);
+      return NewStepFile._(
+        import,
+        filename,
+        package,
+        line,
+        testerType,
+        testerName,
+      );
     }
 
     final pathToTestFolder = p.relative(testFolderName, from: featureDir);
@@ -41,19 +58,35 @@ abstract class StepFile {
         .join(pathToTestFolder, generatorOptions.stepFolder, file)
         .replaceAll(r'\', '/');
     final filename = p.join(testFolderName, generatorOptions.stepFolder, file);
-    return NewStepFile._(import, filename, package, line);
+    return NewStepFile._(
+      import,
+      filename,
+      package,
+      line,
+      testerType,
+      testerName,
+    );
   }
 }
 
 class NewStepFile extends StepFile {
-  const NewStepFile._(super.import, this.filename, this.package, this.line)
-      : super._();
+  const NewStepFile._(
+    super.import,
+    this.filename,
+    this.package,
+    this.line,
+    this.testerType,
+    this.testerName,
+  ) : super._();
 
   final String package;
   final String line;
   final String filename;
+  final String testerType;
+  final String testerName;
 
-  String get dartContent => generateStepDart(package, line);
+  String get dartContent =>
+      generateStepDart(package, line, testerType, testerName);
 }
 
 class ExistingStepFile extends StepFile {